Set docutils>=0.20.0 along with other packages
Apparently the version available in my dev is env(0.21.1) is higher
than what's available to the gitlab-runner. The same is true for
other packages.

Set another package to a previous version to accomodate gitlab runner

Set another package to a previous version to accomodate gitlab runner

Set another package to a previous version to accomodate gitlab runner

Set another package to a previous version to accomodate gitlab runner